feat: added a basic terraform file and setup instructions (#20)
This should help new users be able to scale into AWS a little faster. I left the `.terraform` folder out of the `.gitignore` because it may be useful for teams/people to check that in for state management purposes.

I was unable to properly figure out how to transfer the `yml` and the `pem` file via terraform. However, terraform takes CloudFormation scripts etc, so I'm sure it could be expanded upon.

Please let me know if there are any questions etc. around this process. I'm happy to add more comments to clear some of it up if it is unclear by default.

Also, to give credit where credit is due, this is based heavily on https://github.com/circleci/enterprise-setup at a high level. I made some modifications and sanatized it so it would work for this basic use case. I also referred to the terraform docs.