feat: allow creating or using service accounts
The update in CE is designed to be backward compatible with the existing
functionality, but using existingName should be preferred as this allows
the Helm chart to manage creation of the SA where needed.

This change allows Renovate pods to be assigned roles with IAM
permissions, enabling access to things like private AWS ECR
repositories.